Add handling for query params from both URI and request params
Including non-array params in the signed request object was causing
the following error:

> TypeError: ws.uri.queryParams[key].sort is not a function

This fixes that and adds a test to make sure it's fixed. As I am
not familiar with the tool used to generate test data, I cloned and
modified existing test data.

issue: https://github.com/danieljoos/aws-sign-web/issues/20